requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.12.3
bitsandbytes==0.41.0
cachetools==5.3.1
certifi==2025.6.15
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
deepspeed==0.9.5
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.1.3
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gradio==3.35.2
gradio_client==0.2.9
greenlet==3.2.3
gspread==6.2.1
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.0
huggingface-hub==0.18.0
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
linkify-it-py==2.0.2
-e git+https://github.com/haotian-liu/LLaVA.git@7775b12d6b20cd69089be7a18ea02615a59621cd#egg=llava
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==2.1.3
matplotlib==3.8.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
ninja==1.11.1.1
numpy==1.26.0
oauthlib==3.2.2
orjson==3.9.8
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
pathtools==0.1.2
peft==0.4.0
Pillow==10.0.1
playwright==1.53.0
protobuf==4.24.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.13
pydub==0.25.1
pyee==13.0.0
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.1.1
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.4
rsa==4.9
safetensors==0.4.0
scikit-learn==1.2.2
scipy==1.11.3
selenium==4.34.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
threadpoolctl==3.2.0
TikTokApi==7.1.0
timm==0.6.13
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.4.0
uvicorn==0.23.2
wandb==0.15.12
wavedrom==2.0.3.post3
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
